requirements.txt
youtube_transcript_api
openai
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
pydantic
colorama
starlette